调整块的子结构key

+++ src/block/partial/operate.ts
@@ -3,7 +3,7 @@ import { Block } from "..";
 import { util } from "../../../util/util";
 import { OperatorDirective } from "../../history/declare";
 import { DropDirection } from "../../kit/handle/direction";
-import { BlockUrlConstant } from "../constant";
+import { BlockChildKey, BlockUrlConstant } from "../constant";
 import { Col } from "../element/col";
 import { BlockRenderRange } from "../enum";
 import lodash from 'lodash';
@@ -126,13 +126,13 @@ export class Block$Operator {
     async insertBefore(this: Block, to: Block, childsKey?: string) {
         await to.parent.append(this,
             to.at,
-            childsKey || to.parent.childKey
+            childsKey
         );
     }
     async insertAfter(this: Block, to: Block, childsKey?: string) {
         await to.parent.append(this,
             to.at + 1,
-            childsKey || to.parent.childKey
+            childsKey
         );
     }
     /**
@@ -159,7 +159,14 @@ export class Block$Operator {
         }
     }
     async append(this: Block, block: Block, at?: number, childKey?: string) {
-        if (typeof childKey == 'undefined') childKey = 'childs';
+        if (typeof childKey == 'undefined') {
+            if (block.isLine)
+                childKey = BlockChildKey.childs;
+            else {
+                if (this.hasSubChilds) childKey = BlockChildKey.subChilds
+                else childKey = BlockChildKey.childs;
+            }
+        }
         if (!this.allBlockKeys.some(s => s == childKey)) {
             throw 'not append the block support subkey';
         }
@@ -179,17 +186,17 @@ export class Block$Operator {
         this.page.addBlockUpdate(this);
     }
     async appendBlock(this: Block, blockData: Record<string, any>, at?: number, childKey?: string) {
-        if (typeof childKey == 'undefined') childKey = 'childs';
-        if (!this.allBlockKeys.some(s => s == childKey)) {
-            throw 'not append the block support subkey';
-        }
-        var bs = this.blocks[childKey];
-        if (typeof at == 'undefined') at = bs.length;
         var newBlock = await this.page.createBlock(blockData.url, blockData, this, at, childKey);
         return newBlock;
     }
     async appendArray(this: Block, blocks: Block[], at?: number, childKey?: string) {
-        if (typeof childKey == 'undefined') childKey = 'childs';
+        if (typeof childKey == 'undefined') {
+            if (blocks.first().isLine) childKey = BlockChildKey.childs;
+            else {
+                if (this.hasSubChilds) childKey = BlockChildKey.subChilds;
+                else childKey = BlockChildKey.childs;
+            }
+        }
         var bs = this.blocks[childKey];
         if (typeof at == 'undefined') at = bs.length;
         var b;
@@ -205,8 +212,15 @@ export class Block$Operator {
             }
         }
     }
-    async appendArrayBlockData(this: Block, blocks: Record<string, any>[], at?: number, childKey?: string) {
-        if (typeof childKey == 'undefined') childKey = 'childs';
+    /**
+     * childKey不能为空
+     * @param this 
+     * @param blocks 
+     * @param at 
+     * @param childKey 
+     * @returns 
+     */
+    async appendArrayBlockData(this: Block, blocks: Record<string, any>[], at: number, childKey: string) {
         if (!this.allBlockKeys.some(s => s == childKey)) {
             throw 'not append the block support subkey';
         }
@@ -228,94 +242,6 @@ export class Block$Operator {
         return cs;
     }
     /**
-     * 注意元素移到to元素下面，并非简单的append，
-     * 例如 将一个元素移到一个并排的元素下面时，需要主动的创建一个col，如果当前元素没有col容器时
-     * @param to 
-     * @param direction 
-     */
-    async move(this: Block, to: Block, direction: DropDirection) {
-        var self = this;
-        switch (direction) {
-            case DropDirection.bottom:
-            case DropDirection.top:
-                var row = to.closest(x => x.isBlock);
-                var childsKey = 'childs';
-                if (row.parent.url == BlockUrlConstant.List) {
-                    childsKey = 'subChilds';
-                }
-                if (direction == DropDirection.bottom) {
-                    await row.parent.append(this, row.at + 1, childsKey);
-                }
-                else {
-                    await row.parent.append(this, row.at, childsKey);
-                }
-                break;
-            case DropDirection.left:
-                if (to.isCol || to.parent.isCol) {
-                    var col = to.isCol ? to : to.parent;
-                    var sum = col.childs.sum(x => (x as Col).widthPercent);
-                    var r = Math.round(100 / (col.childs.length + 1));
-                    await col.childs.eachAsync(async (c) => {
-                        await c.updateProps({ widthPercent: ((c as Col).widthPercent / sum) * (100 - r) })
-                    })
-                    var newCol = await this.page.createBlock(BlockUrlConstant.Col, { widthPercent: r }, col.parent, col.at);
-                    await newCol.append(this);
-                }
-                else {
-                    var newRow = await this.page.createBlock(BlockUrlConstant.Row, {
-                        blocks: {
-                            childs: [
-                                { url: BlockUrlConstant.Col, widthPercent: 50 },
-                                { url: BlockUrlConstant.Col, widthPercent: 50 }
-                            ]
-                        }
-                    }, to.parent, to.at)
-                    await newRow.childs.first().append(this);
-                    await newRow.childs.last().append(to);
-                }
-                /**
-                 * 这里新增一个元素，需要调整当前行内的所有元素比例
-                 */
-                break;
-            case DropDirection.right:
-                if (to.isCol || to.parent.isCol) {
-                    var col = to.isCol ? to : to.parent;
-                    var sum = col.childs.sum(x => (x as Col).widthPercent);
-                    var r = Math.round(100 / (col.childs.length + 1));
-                    await col.childs.eachAsync(async (c) => {
-                        await c.updateProps({ widthPercent: ((c as Col).widthPercent / sum) * (100 - r) })
-                    })
-                    var newCol = await this.page.createBlock(BlockUrlConstant.Col, { widthPercent: r }, col.parent, col.at + 1);
-                    await newCol.append(this);
-                }
-                else {
-                    var newRow = await this.page.createBlock(BlockUrlConstant.Row, {
-                        blocks: {
-                            childs: [
-                                { url: BlockUrlConstant.Col, widthPercent: 50 },
-                                { url: BlockUrlConstant.Col, widthPercent: 50 }
-                            ]
-                        }
-                    }, to.parent, to.at)
-                    await newRow.childs.first().append(to);
-                    await newRow.childs.last().append(this);
-                }
-                /**
-                * 这里新增一个元素，需要调整当前行内的所有元素比例
-                */
-                break;
-            case DropDirection.inner:
-                /**这时判断是否可以允许换行的block，还是替换 */
-                var old = to;
-                await to.parent.append(this, old.at);
-                await old.delete();
-                break;
-            case DropDirection.sub:
-                await to.acceptSubFromMove(this);
-                break;
-        }
-    }
-    /**
      * 将一堆blocks拖到this block中
      * @param this 
      * @param blocks 
@@ -327,7 +253,7 @@ export class Block$Operator {
             case DropDirection.bottom:
             case DropDirection.top:
                 var row = this.closest(x => !x.isLine);
-                var childsKey = row.parent.childKey;
+                var childsKey = row.parent.hasSubChilds ? BlockChildKey.subChilds : BlockChildKey.childs;
                 if (direction == DropDirection.bottom) {
                     await row.parent.appendArray(blocks, row.at + 1, childsKey);
                 }
@@ -404,13 +330,14 @@ export class Block$Operator {
                 var cs = this.childs.map(c => c);
                 await this.appendArray(blocks);
                 await cs.eachAsync(async (c) => {
-                    await c.delete()
+                    if (c.isContentEmpty) await c.delete()
                 })
                 break;
             case DropDirection.sub:
-                for (let i = blocks.length - 1; i >= 0; i--) {
-                    await this.acceptSubFromMove(blocks[i]);
+                if (this.hasSubChilds) {
+                    await this.appendArray(blocks, 0, BlockChildKey.subChilds)
                 }
+                else await this.parent.appendArray(blocks, this.at, this.parent.hasSubChilds ? BlockChildKey.subChilds : BlockChildKey.childs)
                 break;
         }
     }
@@ -420,7 +347,7 @@ export class Block$Operator {
             case DropDirection.bottom:
             case DropDirection.top:
                 var row = this.closest(x => x.isBlock);
-                var childsKey = row.parent.childKey;
+                var childsKey = row.parent.hasSubChilds ? BlockChildKey.subChilds : BlockChildKey.childs;
                 if (direction == DropDirection.bottom) {
                     await row.parent.appendArrayBlockData(blocks, row.at + 1, childsKey);
                 }
@@ -495,23 +422,17 @@ export class Block$Operator {
                  * 这时判断是否可以允许换行的block，还是替换 
                  * */
                 var cs = this.childs.map(c => c);
-                await this.appendArrayBlockData(blocks);
+                await this.appendArrayBlockData(blocks, undefined, BlockChildKey.childs);
                 await cs.eachAsync(async (c) => {
-                    await c.delete()
+                    if (c.isContentEmpty)
+                        await c.delete()
                 })
                 break;
             case DropDirection.sub:
-                await this.appendArrayBlockData(blocks, 0, this.childKey);
+                if (this.hasSubChilds) await this.appendArrayBlockData(blocks, 0, this.hasSubChilds ? BlockChildKey.subChilds : BlockChildKey.childs);
+                else await this.parent.appendArrayBlockData(blocks, this.at, this.parent.hasSubChilds ? BlockChildKey.subChilds : BlockChildKey.childs);
                 break;
         }
-    }
-    /**
-     * 表示当前元素如何接收该元素至sub,
-     * @param this 
-     * @param sub  子元素是要移动的
-     */
-    async acceptSubFromMove(this: Block, sub: Block) {
-
     }
     async updateAppear(this: Block, appear: AppearAnchor, newValue: string, range = BlockRenderRange.none) {
         await this.updateProps({ [appear.prop]: newValue });
